requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
